API Bump
LibCustomMenu is now optional, and the right click to craft feature cannot be enabled if LCM is not running.
Anniversary boxes will no longer be attempted to be opened.
Anniversary loot notifications have been added
Will now close the looting window if there are transmute stones that are not lootable because they would put the user over the maximum amount of transmute stones
Added reward handling... not working atm
Added a dismissable roadblock - this is more as a warning when something is wrong but not so wrong as to prevent normal operation
Added a feedback option which is '1g/writ completed
Removed the maze buttons