API Bump
If the player's inventory is full, and you need to withdraw an item for a writ, you will now be able to use the bank without getting kicked out of the bank
Now checks if there is room in the bank prior to trying to deposit items
Fixed a bug with enchanting writs where the addon would not auto exit the enchanting station if the 'acquire a rune' step was incomplete.
Added Ornate gear for writ reward handling, so you can now junk, deposit, or deconstruct ornate gear from writs.
Fixed a bug where the 'quest accept' center screen announcements were always blocked, even if the player had quest announcements turned on
Changed the loot reward handling to no longer rely on translated strings. Instead, it now uses a list of the item IDs for the different possible writ rewards. (Thanks @MelanAster)
In conjunction with the above, have removed a number of no longer used tables from the localization files, and updated the translation guidelines. Notably, strings which are only used for master writs are now marked as optional translations.
Removed GetItemIDFromLink, since ZOS has added GetItemLinkItemId since ages ago
Added the option to deconstruct intricate items that you receive from writs. It will queue the item immediately, but if you logout before it is deconstructed, it will re-queue the item for deconstruction once you login. This way, you don't need to re-visit stations to deconstruct and it will instead deconstruct them when you do writs the next day.