Add support for subtractive runes and the Reduce Spell Harm glyph
The Stirk Fellowship, facing an army of necromancers, put out the
call for crafting Reduce Spell Harm glyphs.  This requires picking
between the additive and subtractive runes for each level.

glyphIds now has an extra field containing +1 or -1 indicating that
an additive or subtractive rune is needed.  itemLinkLevels now contains
the IDs of both additive and subtractive runes for each level.