Report "Could not determine which glyphs to use" if either is missing
While working on enchantSearch, I was getting "Glyph could not be
crafted.  You do not have any <blank> or <blank>." errors.  This is a
bit clearer, as if either rune is unknown, we don't know how to make
the necessary glyph.

In reality, enchantSearch returns nil if it couldn't find a rune
combination, so both fields will be nil anyway, but...still, easier
to debug if you get the proper error message.