Craft multiple glyphs as needed and support Armorer Reistaff quests
Daily crafting writs require crafting a single glyph.  However, the
new Stirk Fellowship event dailies require crafting two glyphs.  So
we can't assume a quantity of 1.

Annoyingly, the quest organization seems to differ as well:

Regular writs...have GetJournalQuestNumConditions() == 3.  The first
step seems to have cur = 0 and max = 1, and be the one to look at.  But
it has QUEST_CONDITION_TYPE_ADVANCE_COMPLETABLE_SIBLINGS (45) as the
conditionType too, despite being the only step with a max quantity.

Stirk Fellowship event dailies have NumConditions() == 2.  They require
crafting two glyphs (instead of 1), and don't have an "acquire a rune"
step at the end.  The /second/ step is the one for crafting the glyphs
with max = 2.  The first step again has ADVANCE_COMPLETABLE_SIBLINGS as
the type but with max = 0, and no actual work to do.  The text of step
1 changes to "Return to Armorer Reistalf" when done.

With these changes, regular writs complete as normal.  The new Stirk
writs craft successfully, then instead of saying "Writ Complete" and
closing, flags "Could not determine which glyphs to use".  Still,
automatically crafting the glyphs is an improvement.